+++ extension/src/views/registryView.ts
@@ -200,7 +200,7 @@ class RecommendedProvider implements TreeDataProvider<Element>, Disposable {
         const extensions: Package[] = [];
 
         for (const pkg of await this.registryProvider.getUniquePackages()) {
-            if (recommendedExtensions.has(pkg.extensionId) && !pkg.isInstalled) {
+            if (recommendedExtensions.has(pkg.extensionId)) {
                 extensions.push(pkg);
             }
         }